% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_country_analysis.R
\name{within_country_analysis}
\alias{within_country_analysis}
\title{Within country analysis with replicate weights and/or plausible values}
\usage{
within_country_analysis(
  cur_country,
  country_var_name,
  cur_variables,
  cur_pv = NULL,
  cur_pv_name = NULL,
  cur_func,
  main_weight,
  rep_weights,
  rep_method,
  rho = NULL,
  dat
)
}
\arguments{
\item{cur_country}{Character vector of length one. Must be the same as in the dataset.}

\item{country_var_name}{Character vector of length one. Name of column containing country names.}

\item{cur_variables}{Character vector. Any column name in the dataset that is not a plausible value used in the analysis function.}

\item{cur_pv}{Character vector. Column names of the current plausible values.}

\item{cur_pv_name}{Character vector of length one. The name the plausible value will be referenced by in the analysis function.}

\item{cur_func}{Analysis function}

\item{main_weight}{Character of length one. Name of the column with main weights.}

\item{rep_weights}{Dataframe with replicate weights}

\item{rep_method}{Character vector of length one. "Fay" or "JK2" according to data}

\item{rho}{# Numeric of length one. Supplied by data provider.}

\item{dat}{Data}
}
\value{
Dataframe with number of rows equal to computed statistics and the follwing columns.
}
\description{
Within country analysis with replicate weights and/or plausible values
}
\author{
Matthias Roth
}
