% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvrep.R
\name{pvrep}
\alias{pvrep}
\title{Within country analysis with replicate weights and/or plausible values using the `survey` and `mitools` packages}
\usage{
pvrep(
  cur_country,
  country_var_name,
  cur_variables,
  cur_pv = NULL,
  cur_pv_name = NULL,
  cur_func,
  main_weight,
  rep_weights,
  rep_method,
  rho = NULL,
  number_of_rep_weights = NULL,
  dat
)
}
\arguments{
\item{cur_country}{Character vector of length one. Must be the same as in the dataset.}

\item{country_var_name}{Character vector of length one. Name of column containing country names.}

\item{cur_variables}{Character vector. Any column name in the dataset that is not a plausible value used in the analysis function.}

\item{cur_pv}{List of character vectors or one character vector. Column names of the current plausible values.}

\item{cur_pv_name}{Character vector. The names the plausible values will be referenced by in the analysis function.}

\item{cur_func}{Analysis function. See Details for information on how to define the analysis function.}

\item{main_weight}{Character of length one. Name of the column with main weights.}

\item{rep_weights}{Character vector of the names of the replicate weights column names in the dataset.}

\item{rep_method}{Character vector of length one. "Fay" or "JK2" according to data}

\item{rho}{Numeric of length one. Supplied by data provider.}

\item{dat}{Data}
}
\value{
Dataframe with number of rows equal to computed statistics and the following columns.
* parameter: Name of the parameter.
* est: Point estimate of the parameter.
* se: Standard error of the parameter.
* df: Degrees of freedom according to Lumely (LINK TO BLOGPOST)
*ci_lwr: Lower bound of the 95% confidence interval.
*ci_upr: Upper bound of the 95% confidence interval.
*p_value: p value of a hypothesis test with H0 = 0.
}
\description{
Currently tailored towards the PIAAC Cycle 2 dataset. Might be expanded in future.
The aim of the function is to make it simpler to write analysis code that works
in the replicate weights/plausible value framework. The function abstracts the loops
over replicate weights/plausible values and pools the results in an appropriate manner.
}
\details{
The analysis function provided to pvrep performs the analysis taking into account
replicate weights or replicate weights & plausible values.

If only replicate weights are required (i.e. the analysis does not involve plausible values)
the function needs the following arguments in this order: data, current_weight, current_variables.
* `data` is the dataframe provided to pvrep.
* `current_weight` is the weight used for the analysis.
* `current_variables` are any other variable names (i.e. provided as a character vector) used in the analysis.
This can be useful for performing loops over different combination of variables)
}
\author{
Matthias Roth
}
